[Node]: replace firewalld approach to set ovirthostnode as default zone
systemctl status firewalld always return 0 in anaconda/kickstart so it's
not possible to determine if we should use firewalld-cmd or
firewalld-offline-cmd. As it's supported as well
the manual installation of release pkgs via dnf/yum this patch
will copy current firewalld.conf for later restore and use
sed cmd to update the zone.

For more info:
Running in chroot, ignoring request -- exit code 0
https://bugzilla.redhat.com/show_bug.cgi?id=1110675

Change-Id: Idef65e4125044fee920c7865203d329ce90efb6f
Signed-off-by: Douglas Schilling Landgraf <dougsland@redhat.com>