node: Node release does not depend on the ovirt-release
Previously the ovirt-release-host-node package required the
same VR of the ovirt-release package, but this dependency is not
necessary, because: It is a valid use case to use the ovirt-release
-host-node rpm from master, but use the ovirt-release file from the
stable branch.

With the dependency it is not possible to have different VR for the
ovirt-release and ovirt-release-host-node packages.
With this patch it is now possible to have different VRs.

Change-Id: Ib3edf5d00c2cfc3b0596359246349ede1a9123f8
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>