deps: gdeploy: changed copr repository
Ramesh left Red Hat and will not be maintaining his repository any longer.
Sac is the main developer of Gdeploy and is working on getting the package
in Fedora before it can get included in the CentOS Storage SIG.
Until the packages are in the repositories from the distributions,
the COPR at https://copr.fedorainfracloud.org/coprs/sac/gdeploy/
contains the latest builds.

Change-Id: I07f08892b3f3ffa7f9719e99993c2a7b199ac81f
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>