el7: epel: add nbdkit
We are adding another mechanism for importing VMs. It uses VDDK,
a proprietary library from VMware. For this to work user has to accept
the license and download the library. Then user has to build a VDDK
plugin for nbdkit with this library. When this is done virt-v2v can be
instructed to make use of nbdkit+VDDK for imports.

Change-Id: I335826729dc728dd7471a0221dc36f7d55c3acac
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>