el8: add NetworkManager from copr
NetworkManager-1.20 is required by nmstate, which is a vdsm dependency,
for oVirt-4.4.

This patch only updates the release for the x86_64 architecture, since
the NetworkManager COPR does not have builds for ppc architecture, and
CentOS 8.1 (which is scheduled for November) is going to ship
NetworkManager-1.20.

Change-Id: I18f701d80eff87f0728bf9d83248ef1fd21c58aa
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>
(cherry picked from commit 9db76387fa1ce377bbd82e8ec1b27311358e17b3)