packaging: Use shorter versions in OS Release "pretty name"
Ansible OS version detection logic will try to get the most
"precise" OS version.

It checks several places in the system for versions, and
compares the versions it finds in these places, to see what
provides the most specific version - evaluated by the number
of numbers in the version (actually the number of dots).

One of these places seems to be the os-release's PRETTY_NAME,
in which we normally put a string of "oVirt Node", together
with the oVirt full release version, i.e.  4.4.1.

The latest release, 4.4.1.1 made the Ansible prefer it over
the CentOS release version - 8.2.2004, since it contains 4
numbers vs. 3.

Thus in order to not break the Ansible OS version detection
logic, we need to make sure that we never provide anything
more specific than the OS release version.

This patch fixes this, so only 2 numbers, i.e. 4.4, will be
provided for the version in the os-release's PRETTY_NAME.

Change-Id: Ibb6ae5da32baf7d9de2005e70d257021514a69ae
Signed-off-by: Lev Veyde <lveyde@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1858234