Use dedicated mirrorlist URL
Currently mirrorlist is served by the resources.ovirt.org repo server
which is not IPv6 capable. Moving to a dedicated URL allows for easier
control over where it's hosted and make the script independent of the main repo.

Currently the URL is already hosted on an IPv6-enabled system.

Sample URL to test on existing deployments:
    https://mirrorlist.ovirt.org/mirrorlist-ovirt-4.4-el8

Change-Id: I6e144c9636984c687f048371f8673fbac7967dc8
Jira-ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-2983
Signed-off-by: Evgheni Dereveanchin <ederevea@redhat.com>