spec: Don't use rpm conditionals during installation time
The changed constructs don't work for our use case - we want an el8
version of the release package to work properly on el9, but the '%if'
conditionals make the entoured section removed for the el8 version.

Let's change the code to use bash conditionals based on values from
'/etc/os-release'.

Signed-off-by: Marcin Sobczyk <msobczyk@redhat.com>
Change-Id: Ie8dcd500d74b16b4b0679193febaf7304f70f17b